+++ src/backend/drive/settings.py
@@ -231,9 +231,6 @@ class Base(Configuration):
         "image/svg+xml",
     ]
 
-    # item versions
-    item_VERSIONS_PAGE_SIZE = 50
-
     # Internationalization
     # https://docs.djangoproject.com/en/3.1/topics/i18n/