+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- ✨(backend) search endpoint for ItemViewSet #312
+
 ## [v0.3.0] - 2025-08-25
 
 ### Added
@@ -22,6 +26,7 @@ and this project adheres to
 - 🐛(front) fix workspace link react cache #310
 - 🐛(backend) allow item partial update without modifying title #316
 - 🌐(front) fix share item wording #315
+
 
 ## [v0.2.0] - 2025-08-18